waifu2x

+++ commands/fun/say.ts
@@ -8,6 +8,14 @@ export default class Say extends Command {
     constructor(discord: Kisaragi, message: Message) {
         super(discord, message, {
             description: "Posts your message.",
+            help:
+            `
+            \`say text\` - Posts the text.
+            `,
+            examples:
+            `
+            \`=>say I love you\`
+            `,
             aliases: [],
             cooldown: 3
         })
@@ -19,6 +27,7 @@ export default class Say extends Command {
         const prefix = await SQLQuery.fetchPrefix(message)
 
         const rawText = Functions.combineArgs(args, 1)
+        if (!rawText) return message.reply("You did not provide any text.")
         await message.channel.send(Functions.checkChar(rawText, 2000, "."))
         if (message.content.startsWith(prefix[0])) await message.delete()
         return
